refactor: student view problems list
