+++ frontend/webreactjs/basewebfrontend/src/component/education/programmingcontestFE/StudentViewProblemList.js
@@ -1,9 +1,40 @@
-import { Link, useHistory } from "react-router-dom";
-import React from "react";
 import MaterialTable from "material-table";
-export default function StudentViewProblemList(props) {
-  const problems = props.problems;
-  const contestId = props.contestId;
+import React, { useEffect, useState } from "react";
+import { Link, useParams } from "react-router-dom";
+import { request } from "./Request";
+export default function StudentViewProblemList() {
+  const { contestId } = useParams();
+  const [problems, setProblems] = useState([]);
+
+  function getContestDetail() {
+    request(
+      "get",
+      "/get-contest-detail-solving/" + contestId,
+      (res) => {
+        setProblems(res.data.list);
+        for (let i = 0; i < res.data.list.length; i++) {
+          let idSource =
+            contestId + "-" + res.data.list[i].problemId + "-source";
+          let tmpSource = localStorage.getItem(idSource);
+          let idLanguage =
+            contestId + "-" + res.data.list[i].problemId + "-language";
+          let tmpLanguage = localStorage.getItem(idLanguage);
+          if (tmpSource == null) {
+            localStorage.setItem(idSource, "");
+          }
+          if (tmpLanguage == null) {
+            localStorage.setItem(idLanguage, "CPP");
+          }
+        }
+      },
+      {}
+    );
+  }
+
+  useEffect(() => {
+    getContestDetail();
+  }, []);
+
   const columns = [
     {
       title: "ProblemID",
@@ -28,7 +59,7 @@ export default function StudentViewProblemList(props) {
   ];
   return (
     <div>
-      <h1> StudentViewProblemList</h1>
+      <h1>Problems List</h1>
       <MaterialTable columns={columns} data={problems} />
     </div>
   );
